- Support for canvas root element ( eg: bpmn diagram )- - Refactor BPMN shapes for using proxies of the basic set ones. - Added shape and proxy for connectors. - Added Basic Connector into Basic Set. - Refactor of SeqFlowShape by use of the connector proxy. - Added canvas grid & toolbar button for swtich grid. - Improved perspective and screens. - Bug fixing.

+++ wirez-core/wirez-commons/wirez-client-common/src/main/java/org/wirez/core/client/canvas/AbstractCanvasHandler.java
@@ -197,7 +197,7 @@ public abstract class AbstractCanvasHandler<D extends Diagram, C extends Abstrac
 
         // Start processing.
         fireProcessingStarted();
-        
+
         treeWalkTraverseProcessor
                 .usePolicy(TreeWalkTraverseProcessor.TraversePolicy.VISIT_EDGE_AFTER_TARGET_NODE)
                 .traverse(diagram.getGraph(), new AbstractTreeTraverseCallback<Graph, Node, Edge>() {
@@ -211,15 +211,21 @@ public abstract class AbstractCanvasHandler<D extends Diagram, C extends Abstrac
             public boolean startNodeTraversal(final Node node) {
                 
                 if ( node.getContent() instanceof View ) {
-                    final View viewContent = (View) node.getContent();
-                    final ShapeFactory factory = 
-                            shapeManager.getFactory( getDefinitionId( viewContent.getDefinition() ) );
 
-                    // Add the node shape into the canvas.
-                    register( factory, node );
-                    applyElementMutation(node, MutationContext.STATIC);
-                    
+                    if ( !isCanvasRoot( node ) ) {
+
+                        final View viewContent = (View) node.getContent();
+                        final ShapeFactory factory =
+                                shapeManager.getFactory( getDefinitionId( viewContent.getDefinition() ) );
+
+                        // Add the node shape into the canvas.
+                        register( factory, node );
+                        applyElementMutation(node, MutationContext.STATIC);
+
+                    }
+
                     return true;
+
                 }
                 
                 return false;
@@ -444,31 +450,64 @@ public abstract class AbstractCanvasHandler<D extends Diagram, C extends Abstrac
     }
     
     public void addChild(final Element parent, final Element child) {
-        final Shape parentShape = canvas.getShape(parent.getUUID());
-        final Shape childShape = canvas.getShape(child.getUUID());
-        handleParentChildZIndex( parent, child, parentShape, childShape, true );
-        canvas.addChildShape(parentShape, childShape);
+
+        if ( !isCanvasRoot( parent ) ) {
+
+            final Shape parentShape = canvas.getShape(parent.getUUID());
+            final Shape childShape = canvas.getShape(child.getUUID());
+            handleParentChildZIndex( parent, child, parentShape, childShape, true );
+            canvas.addChildShape(parentShape, childShape);
+
+        }
+
     }
 
     public void removeChild(final String parentUUID, final String childUUID) {
-        final Shape parentShape = canvas.getShape( parentUUID );
-        final Shape childShape = canvas.getShape( childUUID );
-        handleParentChildZIndex( null, null, parentShape, childShape, false );
-        canvas.deleteChildShape(parentShape, childShape);
+
+        if ( !isCanvasRoot( parentUUID ) ) {
+
+            final Shape parentShape = canvas.getShape( parentUUID );
+            final Shape childShape = canvas.getShape( childUUID );
+            handleParentChildZIndex( null, null, parentShape, childShape, false );
+            canvas.deleteChildShape(parentShape, childShape);
+
+        }
+
+    }
+
+    protected boolean isCanvasRoot( final Element parent ) {
+        return null != parent && isCanvasRoot( parent.getUUID() );
+    }
+
+    protected boolean isCanvasRoot( final String pUUID ) {
+        final String canvasRoot = getDiagram().getSettings().getCanvasRootUUID();
+        return ( null != canvasRoot && null != pUUID && canvasRoot.equals( pUUID ) );
     }
 
     public void dock(final Element parent, final Element child) {
-        final Shape parentShape = canvas.getShape(parent.getUUID());
-        final Shape childShape = canvas.getShape(child.getUUID());
-        handleParentChildZIndex( parent, child, parentShape, childShape, true );
-        canvas.dock(parentShape, childShape);
+
+        if ( !isCanvasRoot( parent) ) {
+
+            final Shape parentShape = canvas.getShape(parent.getUUID());
+            final Shape childShape = canvas.getShape(child.getUUID());
+            handleParentChildZIndex( parent, child, parentShape, childShape, true );
+            canvas.dock(parentShape, childShape);
+
+        }
+
     }
 
     public void undock(final String parentUUID, final String childUUID) {
-        final Shape parentShape = canvas.getShape( parentUUID );
-        final Shape childShape = canvas.getShape( childUUID );
-        handleParentChildZIndex( null, null, parentShape, childShape, false );
-        canvas.undock(parentShape, childShape);
+
+        if ( !isCanvasRoot( parentUUID ) ) {
+
+            final Shape parentShape = canvas.getShape( parentUUID );
+            final Shape childShape = canvas.getShape( childUUID );
+            handleParentChildZIndex( null, null, parentShape, childShape, false );
+            canvas.undock(parentShape, childShape);
+
+        }
+
     }
     
     protected void handleParentChildZIndex( final Element parent,
